adds different class names for the different controllers. The different controllers are mounted from within Makersbnb
