+++ app.rb
@@ -15,26 +15,47 @@ require './app/controllers/bookings.rb'
 require './app/models/account.rb'
 require './app/models/property.rb'
 
-
 # Would be interesting to think about what the class is doing in the app.rb file. BTW, I guess stuff
 # Can be progressively added to classes.
 class Makersbnb < Sinatra::Base
-  
-  # What should this class do when it is called by rackup?
-  # before, the only thing which app.rb did other than declare methods was:
-  
-  
-  # I think there can be some more config in here
+  use LoginLogout
+  use Properties
+  use Bookings
+  # I'm gonna see what happens if I mount a new controller by using use
+  #  it finds it...
+  # I think it's because it needs to inherit config from Makers BNB
   configure do
-    # set :database, "sqlite3:makersbnb.db.sqlite3"
-    # for some reason the DB didn't look to be working when it was in the config block
     set :app_file, __FILE__
     set :root, File.dirname(__FILE__)
     set :views, Proc.new { File.join(root, "app/views") }
     enable :sessions
-
   end
-  # using configure with a block in this way was copied from the structuring sinatra site
-  
+  # I don't yet have a complet understanding of what config does here - is it the same as in spec helpers?
+end
 
+class LoginLogout < Sinatra::Base
+  configure do
+    set :app_file, __FILE__
+    set :root, File.dirname(__FILE__)
+    set :views, Proc.new { File.join(root, "app/views") }
+    enable :sessions
+  end
+end
+
+class Properties < Sinatra::Base
+  configure do
+    set :app_file, __FILE__
+    set :root, File.dirname(__FILE__)
+    set :views, Proc.new { File.join(root, "app/views") }
+    enable :sessions
+  end
+end
+
+class Bookings < Sinatra::Base
+  configure do
+    set :app_file, __FILE__
+    set :root, File.dirname(__FILE__)
+    set :views, Proc.new { File.join(root, "app/views") }
+    enable :sessions
+  end
 end
