+++ app/controllers/properties.rb
@@ -1,5 +1,5 @@
 
-class Makersbnb < Sinatra::Base
+class Properties < Sinatra::Base
 
 
   get '/properties' do
@@ -14,21 +14,11 @@ class Makersbnb < Sinatra::Base
   end
 
   post '/properties' do
-    # Store property in database
     property = {name: params[:name], description: params[:description], price: params[:price_per_night].to_i, account_id: session[:user_id]}
     Property.create(property)
     redirect '/properties'
   end
   
-
-  def is_logged_in?
-    current_user != nil
-  end
-
-  def current_user
-    session[:user_id]
-  end
-
-  # run! if app_file == $0
 end
+
